+++ trunk/RenderSystems/Direct3D7/include/OgreD3D7RenderSystem.h
@@ -219,7 +219,7 @@ namespace Ogre {
         /** See
           RenderSystem
          */
-        void convertColourValue(const ColourValue& colour, unsigned long* pDest);
+        void convertColourValue(const ColourValue& colour, uint32* pDest);
 
         // -----------------------------
         // Low-level overridden members
